requirements.txt
accelerate
datasets
editdistance
hydra-core
ipdb
matplotlib
nltk
omegaconf
ordered-set
ply
scikit-learn
tensorboard
torch
transformers
wandb
deepspeed
seaborn
torchvision
